Split out `camera_init` vs. `camera_capture`, comments, remove IWDG
