+++ firmware/Core/Inc/camera/camera_capture.h
@@ -1,8 +1,7 @@
-#ifndef INCLUDE_GUARD__CAMERA_COMMANDS_H
-#define INCLUDE_GUARD__CAMERA_COMMANDS_H
+#ifndef INCLUDE_GUARD__CAMERA_CAPTURE_H
+#define INCLUDE_GUARD__CAMERA_CAPTURE_H
 
 #include <stdint.h>
-#include <stdbool.h>
 
 #define CAM_SENTENCE_LEN    67
 
@@ -11,10 +10,7 @@
 
 #define CAM_BYTES_TO_RECEIVE_PER_HALF_CALLBACK ((CAM_SENTENCE_LEN) * (CAM_SENTENCES_PER_HALF_CALLBACK))
 
-uint8_t CAM_change_baudrate(uint32_t bitrate);
 
-uint8_t CAM_setup();
-uint8_t CAM_test();
 void CAM_repeated_error_log_message();
 
 
@@ -30,14 +26,6 @@ typedef enum CAM_capture_status_enum {
 } CAM_capture_status_enum;
 
 
-/**
- * Transmits ASCII telecommand based on input
- * @param lighting - lighting should be a *lower case* char
- * 			d - daylight ambient light
- * 			m - medium ambient light
- * 			n - night ambient light
- * 			s - solar sail contrast and light
- */
 CAM_capture_status_enum CAM_capture_image(char filename_str[], char lighting_mode);
 
-#endif // INCLUDE_GUARD__CAMERA_CAMERA_COMMANDS_H
+#endif // INCLUDE_GUARD__CAMERA_CAPTURE_H
