+++ firmware/Core/Src/telecommands/lfs_telecommand_defs.c
@@ -356,8 +356,10 @@ uint8_t TCMDEXEC_fs_read_text_file(
         DEBUG_uart_print_str((char*)rx_buffer);
         // TODO: send to radio
         DEBUG_uart_print_str("TODO: send data to radio from TCMD_fs_read_text_file()\n");
-        osDelay(300);
-        HAL_IWDG_Refresh(&hiwdg);
+
+        // The following block is to enable printing out camera data. Not permissible in normal operation.
+        // osDelay(300);
+        // HAL_IWDG_Refresh(&hiwdg);
     }
