+++ firmware/Core/Src/uart_handler/uart_handler.c
@@ -2,7 +2,7 @@
 #include "debug_tools/debug_uart.h"
 #include "mpi/mpi_command_handling.h"
 #include "uart_handler/uart_error_tracking.h"
-#include "camera/camera_commands.h"
+#include "camera/camera_capture.h"
 #include "log/log.h"
 
 #include "main.h"
